refactor: src/main/java/seedu.address/commons/events/
