--- src/main/java/seedu/address/commons/events/model/AddressBookChangedEvent.java
@@ -1,19 +1,0 @@
-package seedu.address.commons.events.model;
-
-import seedu.address.commons.events.BaseEvent;
-import seedu.address.model.ReadOnlyAddressBook;
-
-/** Indicates the AddressBook in the model has changed*/
-public class AddressBookChangedEvent extends BaseEvent {
-
-    public final ReadOnlyAddressBook data;
-
-    public AddressBookChangedEvent(ReadOnlyAddressBook data) {
-        this.data = data;
-    }
-
-    @Override
-    public String toString() {
-        return "number of persons " + data.getPersonList().size();
-    }
-}
