--- src/main/java/seedu/address/commons/events/ui/PersonPanelSelectionChangedEvent.java
@@ -1,26 +1,0 @@
-package seedu.address.commons.events.ui;
-
-import seedu.address.commons.events.BaseEvent;
-import seedu.address.model.person.Person;
-
-/**
- * Represents a selection change in the Person List Panel
- */
-public class PersonPanelSelectionChangedEvent extends BaseEvent {
-
-
-    private final Person newSelection;
-
-    public PersonPanelSelectionChangedEvent(Person newSelection) {
-        this.newSelection = newSelection;
-    }
-
-    @Override
-    public String toString() {
-        return getClass().getSimpleName();
-    }
-
-    public Person getNewSelection() {
-        return newSelection;
-    }
-}
