+++ src/main/java/seedu/address/commons/events/ui/TaskPanelSelectionChangedEvent.java
@@ -1,0 +1,26 @@
+package seedu.address.commons.events.ui;
+
+import seedu.address.commons.events.BaseEvent;
+import seedu.address.model.task.Task;
+
+/**
+ * Represents a selection change in the Task List Panel
+ */
+public class TaskPanelSelectionChangedEvent extends BaseEvent {
+
+
+    private final Task newSelection;
+
+    public TaskPanelSelectionChangedEvent(Task newSelection) {
+        this.newSelection = newSelection;
+    }
+
+    @Override
+    public String toString() {
+        return getClass().getSimpleName();
+    }
+
+    public Task getNewSelection() {
+        return newSelection;
+    }
+}
